Remove all mention of Mason from trove-setup
The Mason referred to here is the 1st version of the Mason continuous
delivery tool. There have been no instances of this for two years.
We have made two subsequent Mason implementations since then which don't
require coupling in Trove in order to work.

As well as removing unneeded configuration, this will fix the misleading
warning that users see on Git pushes:

     remote: [git.baserock.org] Notifying Mason of changes...
     remote: [git.baserock.org] Notification failed somehow